A bug fixed in module selection; id="redline" is
Restored because we use it in the code for selection of category.
The stylesheet now contains the styles for the redline, taken from the
inline code.

Module-selection-modified.php is now well-commented and debugged: rows
were opened but never closed. Also, the name of variable capable for the
subcategories was $cat, I switched it to $subcat .

I switched the if($i%$col==0) to if(($i%$col)==0) to ensure the right
order of operations.