put up a StickyTooltip that, when a module is hovered over
shows a box with an image (may be a .gif file) of the module so that the
user loves it. :)

The .css and the .js files are from the plugin.

It works on the module-selection-modified page but seems not passed to
the -art, -junk, -lego, -minecraft pages and it doesn't work there. :(