in the stylesheet: removed all invalid "cursor:hand" +
played a lot with the space for caprions+buttons+links so that it goes
to the right of the slider and doesn't overlap with it.
Placed buttons both on the Instructions and Applications pages in the
bottom + rewrote the code so that they are symmetrical and work the same
way.
Because the slider doesn't overlap with the captions+buttons' div, we
don't need "seems about right" margins on the left. Way better! :)
Also, on the main page, the gray line now is better-set-up to h=28px.
Before, it was 100px, and had margin-top=-72px.
The borders around the images for module of the week and the missions is
thinner and grayer.
The "navigation-bar" is now spelled right.

In the module stylesheet, #slides was overwritten. The code there is
also reindented.

The overview-button is cropped better. :)